Create README.md
Welcome to crash, a simple wrapper for the unix shell using fork() and exec() to run unix commands. 

MUST RUN ON LINUX

In order to run and create the unit tests, follow the setup instructions here: https://www.eriksmistad.no/getting-started-with-google-test-on-ubuntu/

When using cmake, ensure that you are passing the command "-DCMAKE_BUILD_TYPE=Debug" if you wish to build for gdb debugging.